finished the battery

+++ src/at/gabriel/oo/basics/batteries/Battery.java
@@ -20,7 +20,7 @@ public class Battery {
         return state;
     }
 
-    public void setState(int state) {
+    public void setState(double state) {
         this.state = state;
     }
 }
